Revise GEMM lab around tiled optimization and student fusion tasks

+++ src/gemm_lab/utils/correctness.py
@@ -17,13 +17,15 @@ def max_abs_diff(a: torch.Tensor, b: torch.Tensor) -> float:
     return (a - b).abs().max().item()
 
 
-def run_correctness(shape_list: list[tuple[int, int, int]], dtype: torch.dtype, kernel: str) -> list[dict[str, float]]:
+def run_correctness(
+    shape_list: list[tuple[int, int, int]], dtype: torch.dtype, kernel: str
+) -> list[dict[str, float]]:
     rows: list[dict[str, float]] = []
     cfg = GemmConfig(kernel=kernel)  # type: ignore[arg-type]
     for m, n, k in shape_list:
         a = torch.randn((m, k), device="cuda", dtype=dtype)
         b = torch.randn((k, n), device="cuda", dtype=dtype)
-        ref = torch.matmul(a, b).float()
+        ref = torch.matmul(a.float(), b.float()).to(a.dtype)
         out = gemm(a, b, cfg=cfg)
         rows.append({"m": m, "n": n, "k": k, "max_abs": max_abs_diff(out, ref)})
     return rows
